+++ jackson/src/main/scala/com/twitter/finatra/json/internal/serde/ArrayElementsOnNewLinesPrettyPrinter.scala
@@ -5,4 +5,6 @@ import com.fasterxml.jackson.core.util.DefaultPrettyPrinter
 
 private[finatra] object ArrayElementsOnNewLinesPrettyPrinter extends DefaultPrettyPrinter {
   _arrayIndenter = DefaultIndenter.SYSTEM_LINEFEED_INSTANCE
+
+  override def createInstance(): DefaultPrettyPrinter = ArrayElementsOnNewLinesPrettyPrinter
 }